// Frontend-Ventanilla/my-app/src/app/model/RegistroVentaEmpleadoDTO.ts
export class RegistroVentaEmpleadoDTO{

    fechaVenta: string = "";
    horaDeVenta: string = "";
    cantidad: number = 0;
    precioUnitario: number = 0;
    nombreProducto: string = "";
    codigoCliente: number | null = 0;
    codigoEmpleado: number = 0;
    dinero: number = 0;
    //correoCliente: string = "";

}